requirements.txt
PyMuPDF
Pillow
google-generativeai
markdown-it-py